Add useGetCleanBranches hook  + Add filters to useGetFacets
- useGetCleanBranches is a hook that returns an array of clean branches. The hook retrieves an array of branches from the config object, excludes any branches that are blacklisted, and then cleans the IDs of the remaining branches. This is used to filter out blacklisted branches in useGetFacets and useSearchWithPaginationQuery.

- To avoid cycle dependencies, I have restructured the code to remove circular dependencies between modules.
