+++ src/apps/search-result/search-result.tsx
@@ -10,17 +10,11 @@ import {
 } from "../../core/dbc-gateway/generated/graphql";
 import { Work } from "../../core/utils/types/entities";
 import {
-  formatFacetTerms,
+  createFilters,
   useGetFacets
 } from "../../components/facet-browser/helper";
 import useFilterHandler from "./useFilterHandler";
-import { FilterItemTerm, TermOnClickHandler } from "./types";
-import { useConfig } from "../../core/utils/config";
-import { AgencyBranch } from "../../core/fbs/model";
-import {
-  excludeBlacklistedBranches,
-  cleanBranchesId
-} from "../../components/reservation/helper";
+import { TermOnClickHandler } from "./types";
 import { useStatistics } from "../../core/statistics/useStatistics";
 import { useCampaignMatchPOST } from "../../core/dpl-cms/dpl-cms";
 import {
@@ -32,6 +26,7 @@ import FacetBrowserModal from "../../components/facet-browser/FacetBrowserModal"
 import { statistics } from "../../core/statistics/statistics";
 import FacetLine from "../../components/facet-line/FacetLine";
 import { getUrlQueryParam } from "../../core/utils/helpers/url";
+import useGetCleanBranches from "../../core/utils/useGetCleanBranches";
 
 interface SearchResultProps {
   q: string;
@@ -39,18 +34,7 @@ interface SearchResultProps {
 }
 
 const SearchResult: React.FC<SearchResultProps> = ({ q, pageSize }) => {
-  const config = useConfig();
-  const branches = config<AgencyBranch[]>("branchesConfig", {
-    transformer: "jsonParse"
-  });
-  const blacklistBranches = config("blacklistedSearchBranchesConfig", {
-    transformer: "stringToArray"
-  });
-  const whitelistBranches = excludeBlacklistedBranches(
-    branches,
-    blacklistBranches
-  );
-  const cleanBranches = cleanBranchesId(whitelistBranches);
+  const cleanBranches = useGetCleanBranches();
   const [resultItems, setResultItems] = useState<Work[]>([]);
   const [hitcount, setHitCount] = useState<number>(0);
   const [canWeTrackHitcount, setCanWeTrackHitcount] = useState<boolean>(false);
@@ -117,18 +101,6 @@ const SearchResult: React.FC<SearchResultProps> = ({ q, pageSize }) => {
     // We only want to do this once, so we need the dependency array empty
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  const createFilters = (
-    facets: {
-      [key: string]: { [key: string]: FilterItemTerm };
-    },
-    branchIdList: string[]
-  ) => {
-    return {
-      ...formatFacetTerms(facets),
-      ...(cleanBranches ? { branchId: branchIdList } : {})
-    };
-  };
 
   const { data } = useSearchWithPaginationQuery({
     q: { all: q },
